==> notebooks/05_model_job_baseline.ipynb <==
# IT Educational Institute

# Constants
DF_PATH = "../data/processed/2_cleaned_data.pkl"

# Specific Role and Tech coluns from Data
ROLE_COLS = ['DevType']
TECH_COLS = ['LanguageHaveWorkedWith', 'DatabaseHaveWorkedWith',
            'PlatformHaveWorkedWith', 'WebframeHaveWorkedWith',
            'MiscTechHaveWorkedWith', 'ToolsTechHaveWorkedWith',
            'NEWCollabToolsHaveWorkedWith']

MLFLOW_TRACKING_URI = "../models/mlruns"
MLFLOW_EXPERIMENT_NAME = "skills_jobs_stackoverflow"

LOG_PATH = "../models/temp/"
LOG_DATA_PKL = "data.pkl"
LOG_MODEL_PKL = "model.pkl"
LOG_METRICS_PKL = "metrics.pkl"

# Importing libraries
import os
from pathlib import Path
import logging
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly
import pickle

## Function

# Creating a function to save the quality scores of models
def calculate_quality(ground_truth, predictions, metric_function, sort_values=False):
    quality_scores = {}
    for col in predictions.columns:
        role_pred = predictions[col].copy()
        role_truth = ground_truth[col].copy()
        quality_scores[col] = round(metric_function(role_truth, role_pred) * 100, 2)
        
    quality_scores = pd.Series(quality_scores.values(), index=quality_scores.keys())
    if sort_values:
        quality_scores = quality_scores.sort_values()
        
    return quality_scores

## Initialize

### Create Directories

# Creating directories for logging
Path(MLFLOW_TRACKING_URI).mkdir(parents=True, exist_ok=True)
Path(LOG_PATH).mkdir(parents=True, exist_ok=True)

### Read Data

# Reading data
df = pd.read_pickle(DF_PATH)

# Displaying first 5 osbervations
df.head()

### Imbalance Classes

# Checking the total samples of roles
roles_df = df['DevType'].copy()
roles_df.sum().sort_values(ascending=False)

# Resample roles
samples_per_class = 1200
resampled_roles = []

for role_col in roles_df.columns:
    sub_df = roles_df.loc[roles_df[role_col] == 1].copy()
    
    if len(sub_df) < samples_per_class:
        # Upsample
        sub_df = sub_df.sample(samples_per_class, replace=True, random_state=0)
    else:
        # Downsample
        sub_df = sub_df.sample(samples_per_class, random_state=0)
        
    resampled_roles.append(sub_df)

# Concatenating DataFrames
roles_df = pd.concat(resampled_roles)
df = df.loc[roles_df.index].copy()

# Checking the total number of samples after balancing
roles_df.sum().sort_values(ascending=False)

### Train Test Split

# importing train_test_split from sklearn
from sklearn.model_selection import train_test_split

# Creating train and test datasets
X_train, X_test, y_train, y_test = train_test_split(df.drop('DevType',axis=1),
                                                    df['DevType'],
                                                    random_state=0)

# Checking the shapes of both datasets
print("Shape of X_train:", X_train.shape)
print("Shape of X_test:", X_test.shape)
print("Shape of y_train:", y_train.shape)
print("Shape of y_test:", y_test.shape)

## Train Models

### Initialize MLflow

# Importing mlflow and MlflowClient
import mlflow
from mlflow.tracking import MlflowClient

# Initialize client and experiment
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()
mlflow.set_experiment(MLFLOW_EXPERIMENT_NAME)
exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)

### 1. Logistic Regression

# Importing necessary libraries
from sklearn.pipeline import make_pipeline, Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.multioutput import MultiOutputClassifier
from sklearn.linear_model import LogisticRegression

# Creating a pipeline
clf = make_pipeline(StandardScaler(),
                   MultiOutputClassifier(LogisticRegression()))

# Fitting pipeline to train data
clf.fit(X_train.values, y_train.values)

# importing metrics from sklearn
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

# Evaluating on training set
predictions = pd.DataFrame(clf.predict(X_train.values),
                           columns=y_train.columns)
train_scores = {score.__name__: calculate_quality(y_train, predictions, score)
               for score in [accuracy_score, precision_score, recall_score, f1_score]}
train_scores = pd.concat(train_scores, axis=1)

# Evaluating on testing set
predictions = pd.DataFrame(clf.predict(X_test.values), columns=y_train.columns)
test_scores = {score.__name__: calculate_quality(y_test, predictions, score)
               for score in [accuracy_score, precision_score, recall_score, f1_score]}
test_scores = pd.concat(test_scores, axis=1)
mean_test_scores = test_scores.mean()

# Displaying the mean of the test scores
print(mean_test_scores)

# Displaying the test_scores DataFrame
test_scores

## Log Run
### 1. Prepare

# Data details
data_details = {'data_path': DF_PATH,
               'training_indices': X_train.index.tolist(),
               'testing_indices': X_test.index.tolist(),
               'features_names': X_train.columns.droplevel().tolist(),
               'targets_names': y_train.columns.tolist()}

# Writing data in a pickle file
with open(os.path.join(LOG_PATH, LOG_DATA_PKL), 'wb') as output_file:
    pickle.dump(data_details, output_file)

# Model
model = {'model_description': 'Baseline model: Logistic Regression',
        'model_details': str(clf),
        'model_object': clf}

# Writing model data in a pickle file
with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), 'wb') as output_file:
    pickle.dump(model, output_file)

# Metrics
classes_metrics = {'train_scores': train_scores,
                  "test_scores": test_scores}

# Writing metrics to a pickle file
with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), 'wb') as output_file:
    pickle.dump(classes_metrics, output_file)

### 2. Log

# Start a new run and track
with mlflow.start_run(experiment_id=exp.experiment_id, run_name=model['model_description']):
    # Log pickles
    mlflow.log_artifacts(LOG_PATH)
    
    # Track metrics
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)
